requirements.txt
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
Django==5.2.8
django-allauth==65.13.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
idna==3.11
lmstudio==1.5.0
msgspec==0.20.0
packaging==25.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
wsproto==1.3.2
newsapi-python
